+++ src/warcraftTD/hud/MainMenu.java
@@ -1,7 +1,11 @@
 package warcraftTD.hud;
 
+import warcraftTD.World;
+import warcraftTD.WorldEditor;
+import warcraftTD.WorldGame;
 import warcraftTD.libs.StdDraw;
 import warcraftTD.utils.Animation;
+import warcraftTD.utils.Level;
 import warcraftTD.utils.Position;
 
 import javax.sound.sampled.LineUnavailableException;
@@ -10,7 +14,6 @@ import javax.swing.JFileChooser;
 import javax.swing.filechooser.FileNameExtensionFilter;
 import java.io.File;
 
-import java.awt.*;
 import java.io.IOException;
 
 public class MainMenu extends Interface {
@@ -19,10 +22,18 @@ public class MainMenu extends Interface {
     private double delta_time;
     private boolean needReleaseMouse;
     private Animation background;
+    private World nextWorld;
+    private Image title;
+
+    public void setNeedReleaseMouse(boolean needReleaseMouse) {
+        this.needReleaseMouse = needReleaseMouse;
+    }
 
     public MainMenu(int width, int height) throws UnsupportedAudioFileException, IOException, LineUnavailableException {
         StdDraw.setCanvasSize(width, height);
         StdDraw.enableDoubleBuffering();
+
+        this.nextWorld = null;
 
         loadBackground("images/mainMenuBackground/capybara/capybara_");
         this.groupBox = new HorizontalGroupBox(new Position(0.5,0.5), 0.4,0.6,this, "");
@@ -37,7 +48,13 @@ public class MainMenu extends Interface {
         btn = new Button(new Position(0.5,0.1),0.2,0.1,"images/mm_button_quit.png","images/mm_button_quit_hover.png","quit", this);
         this.groupBox.addHUDElement(btn);
 
+        this.title = new Image(new Position(0.5,0.85), 0.8,0.25,this, "images/titlegame.png");
+        this.getListElements().add(this.title);
+
         this.groupBox.ShowBox(0.5,0.0);
+
+        btn = new Button(new Position(0.105,0.055),0.2,0.1,"images/leveleditor_button.png","images/leveleditor_button_hover.png","leveleditor", this);
+        this.getListElements().add(btn);
 
         btn = new Button(new Position(0.9,0.05),0.06,0.08,"images/Capybara_Button.png","images/Capybara_Button_hover.png","switchCapybara", this);
         this.getListElements().add(btn);
@@ -65,7 +82,7 @@ public class MainMenu extends Interface {
     }
 
     @Override
-    public void makeAction(String action, Element from) {
+    public void makeAction(String action, Element from) throws IOException, LineUnavailableException, UnsupportedAudioFileException {
         switch (action) {
             case "quit":
                 System.exit(0);
@@ -80,26 +97,51 @@ public class MainMenu extends Interface {
                 int result = fileChooser.showOpenDialog(StdDraw.getFrame());
                 if (result == JFileChooser.APPROVE_OPTION) {
                     File selectedLvl = fileChooser.getSelectedFile();
+                    Level level = new Level(selectedLvl);
+                    nextWorld = level.getWorld(this);
+                    this.quit = true;
                 }
                 break;
             case "lvl1":
                 this.needReleaseMouse = true;
                 File lvl1 = new File("levels/level1.tdl");
+                Level level = new Level(lvl1);
+                nextWorld = level.getWorld(this);
+                this.quit = true;
                 break;
             case "lvl2":
                 this.needReleaseMouse = true;
                 File lvl2 = new File("levels/level2.tdl");
+                Level level2 = new Level(lvl2);
+                nextWorld = level2.getWorld(this);
+                this.quit = true;
                 break;
             case "switchCapybara":
+                this.needReleaseMouse = true;
                 loadBackground("images/mainMenuBackground/capybara/capybara_");
                 break;
             case "switchCat":
+                this.needReleaseMouse = true;
                 loadBackground("images/mainMenuBackground/kitten/The most dangerous kitten in the world_");
+                break;
+            case "leveleditor":
+                nextWorld = new WorldEditor(1200,800,this);
+                this.quit = true;
                 break;
         }
     }
 
-    public void start() throws UnsupportedAudioFileException, IOException, LineUnavailableException {
+    public void run() throws UnsupportedAudioFileException, IOException, LineUnavailableException {
+        this.quit = false;
+        this.nextWorld = null;
+        while(nextWorld==null){
+            this.startLoop();
+        }
+        this.nextWorld.setNeedReleaseMouse(true);
+        this.nextWorld.run();
+    }
+
+    public void startLoop() throws UnsupportedAudioFileException, IOException, LineUnavailableException {
         while(!quit){
             long time_nano = System.nanoTime();
