+++ src/warcraftTD/towers/projectiles/Bomb.java
@@ -1,6 +1,6 @@
 package warcraftTD.towers.projectiles;
 
-import warcraftTD.World;
+import warcraftTD.WorldGame;
 import warcraftTD.monsters.Monster;
 import warcraftTD.utils.Position;
 import warcraftTD.utils.Vector;
@@ -10,7 +10,7 @@ import java.awt.*;
 public class Bomb extends Projectile {
   double rangeExplosion;
 
-  public Bomb(Position initialPosition, Vector direction, World world, int damage, double rangeExplosion) {
+  public Bomb(Position initialPosition, Vector direction, WorldGame world, int damage, double rangeExplosion) {
     super(initialPosition, direction, world, damage);
     this.sprite = "images/bomb.png";
     this.rotating = false;
